Classification experiments (#1260)
* refactor classification experiments

* remove testing files

* Update test_experiments.py

* docs

* docs

* Update classification_experiments.py

* Update experiments.py

* Update experiments.py

* Delete basic_benchmarking.py

* Update classification_experiments.py

* classification experiments fixes and documentation

* text fix
